refactor: refactor pages directory for easier navigation
