+++ src/lib/pages/book/bookPageUtils.ts
