+++ src/lib/pages/stats/Stats.svelte
