+++ test-italic.html
@@ -22,7 +22,7 @@
 
     <script type="module">
       // Import the renderSentenceHTML function
-      import { renderSentenceHTML } from "./src/lib/pages/bookPageUtils.ts";
+      import { renderSentenceHTML } from "./src/lib/pages/book/bookPageUtils.ts";
 
       const testSentence = {
         en: "There was nothing so _very_ remarkable in that; nor did Alice think it so _very_ much",
